requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.2
jinja2==3.1.5
python-multipart==0.0.20
